Move spine definitions
